# EXPORT APIDOG.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "My Project",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/libro/libros": {
      "put": {
        "summary": "Editar un Libro",
        "x-apidog-folder": "POO2023/Libros",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "POO2023/Libros"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "x-apidog-ignore-properties": [],
                "x-apidog-orders": []
              },
              "example": {
                "id": 1,
                "titulo": "IT",
                "autor": "Stephen King",
                "isbn": "ISBN-10-0451450523",
                "numeroInventario": 554,
                "estado": "DISPONIBLE"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Mensaje": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "Mensaje"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "Mensaje"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/412546/apis/api-4694499-run",
        "security": []
      },
      "post": {
        "summary": "Guardar un Libro",
        "x-apidog-folder": "POO2023/Libros",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "POO2023/Libros"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "x-apidog-ignore-properties": [],
                "x-apidog-orders": []
              },
              "example": {
                "titulo": "La Masacre de Texas",
                "autor": "Stephen King",
                "isbn": "ISBN-10-01502848",
                "numeroInventario": 1,
                "estado": "DISPONIBLE"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Mensaje": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "Mensaje"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "Mensaje"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/412546/apis/api-4694521-run",
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": []
}